requirements.txt
fastapi
boto3
pymongo
uvicorn[standard]
pandas
weaviate-client
moviepy
numpy
-e .
